services/speech_transcriber_gpu/requirements.txt
faster-whisper
google-cloud-firestore
fastapi
uvicorn
ffmpeg-python
torch
numpy
